refactor(control): split reload transaction
