+++ crates/honk-core/src/control/reload/transaction.rs
@@ -1,97 +1,12 @@
 use super::*;
 
 impl ControlPlane {
-    /// Stop and join the prior generation's warm coordinator. Aborting the
-    /// parent drops its JoinSet, so in-flight child dispatches are cancelled
-    /// without becoming health or per-outbound error events.
-    pub(super) async fn stop_udp_warm_coordinator(&self) {
-        let handle = self.udp_warm_task.lock().await.take();
-        if let Some(handle) = handle {
-            handle.abort();
-            let _ = handle.await;
-        }
-    }
-
-    /// Start a warm coordinator bound to one immutable runtime generation.
-    /// A zero count releases the prior UDP retention set without creating a
-    /// task or touching attempt metrics. Positive counts re-rank after every
-    /// probe cycle.
-    pub(super) async fn start_udp_warm_coordinator(
-        &self,
-        generation: Arc<honk_outbound::runtime::OutboundRuntimeRegistry>,
-    ) {
-        if generation.is_shutdown() {
-            return;
-        }
-        let count = self.config.read().await.global.udp_warm_node_count;
-        if count == 0 {
-            reconcile_udp_warm_retention(&[], &generation, &self.stats, &self.udp_warm_ids).await;
-            return;
-        }
-        let connect_timeout = {
-            let config = self.config.read().await;
-            Duration::from_millis(config.global.connect_timeout_ms)
-        };
-        let proxy_registry = self.proxy_registry.clone();
-        let dispatch = Arc::new(move |generation, node_id| {
-            let proxy_registry = proxy_registry.clone();
-            async move {
-                proxy_registry
-                    .warm_udp(generation, node_id, connect_timeout)
-                    .await
-            }
-        });
-        let handle = tokio::spawn(run_udp_warm_coordinator(
-            self.config.clone(),
-            self.group_manager.clone(),
-            generation,
-            self.stats.clone(),
-            dispatch,
-            self.udp_warm_ids.clone(),
-        ));
-        *self.udp_warm_task.lock().await = Some(handle);
-    }
-
-    pub(super) async fn stop_selector_warm_coordinator(&self) {
-        let handle = self.selector_warm_task.lock().await.take();
-        if let Some(handle) = handle {
-            handle.abort();
-            let _ = handle.await;
-        }
-    }
-
-    /// Pin every configured Selector leaf in this immutable runtime
-    /// generation. Choice changes wake the task immediately; the periodic
-    /// pass repairs independently lost sessions and consumed bare sockets.
-    pub(super) async fn start_selector_warm_coordinator(
-        &self,
-        generation: Arc<honk_outbound::runtime::OutboundRuntimeRegistry>,
-    ) {
-        if generation.is_shutdown() {
-            return;
-        }
-        let handle = tokio::spawn(run_selector_warm_coordinator(SelectorWarmCoordinator {
-            config: self.config.clone(),
-            group_manager: self.group_manager.clone(),
-            notify: self.selector_warm_notify.clone(),
-            resources: SelectorWarmResources {
-                generation,
-                proxy_registry: self.proxy_registry.clone(),
-                connection_pool: self.connection_pool.clone(),
-                stats: self.stats.clone(),
-                selected_ids: self.selector_warm_ids.clone(),
-                bare_warm: self.selector_bare_warm.clone(),
-            },
-        }));
-        *self.selector_warm_task.lock().await = Some(handle);
-    }
-
     /// Atomically publish a rebuilt router, config, group manager, outbound
     /// runtime generation, DNS runtime, and exact eBPF routing plan. Build
     /// failures leave the current generation untouched; an eBPF push failure
     /// replays the exact active plan before admission resumes. SIGHUP and
     /// subscription merges share this command-channel-serialized path.
-    pub(super) async fn apply_runtime_config(
+    pub(in crate::control) async fn apply_runtime_config(
         &self,
         new_config: Config,
         drain: &DrainTracker,
@@ -536,23 +451,6 @@ impl ControlPlane {
         }
     }
 
-    /// Merge freshly fetched subscription nodes into the running config,
-    /// replacing the previous node set of `subscription_id`, and run the
-    /// shared rebuild pipeline.
-    ///
-    /// Production callers go through `ControlCommand::MergeSubscription` on
-    /// the command channel (which keeps merges serialized against SIGHUP
-    /// reloads); this public wrapper exists so integration tests can drive a
-    /// merge without binding the TPROXY accept loop.
-    pub async fn merge_subscription_nodes(&self, subscription_id: uuid::Uuid, nodes: Vec<Node>) {
-        let new_config = {
-            let current = self.config.read().await;
-            config_with_subscription_nodes(&current, subscription_id, nodes)
-        };
-        let drain = DrainTracker::new();
-        self.apply_runtime_config(new_config, &drain).await;
-    }
-
     /// Build a DNS forwarder from an explicit config (used by the reload
     /// pipeline's build phase — must not read live state, so the caller can
     /// abort before commit without having mutated anything).
@@ -607,289 +505,5 @@ impl ControlPlane {
             .with_hosts_from_config(&config.dns)?,
         );
         Ok((forwarder, dns_upstream_pool))
-    }
-
-    /// Rebuild the [`GroupManager`] from the current config after a reload.
-    ///
-    /// A fresh manager is installed into the shared cell so every holder
-    /// (control plane, per-connection handles, clash API) picks up new or
-    /// changed groups at once. Runtime selector choices migrate by group
-    /// name (choices whose group or selected node vanished are dropped);
-    /// cache.db-backed choices survive because every change is persisted
-    /// at set time, so no cache.db restore runs here. The alive set's
-    /// health-check registrations and URLTest group table are refreshed to
-    /// match the new group membership, and the node → eBPF outbound id map
-    /// (`outbound_id_map`, already refreshed by the reload path) is built
-    /// from the same config, keeping the two consistent.
-    pub async fn reload_group_manager(&self) {
-        let (groups, nodes) = {
-            let config = self.config.read().await;
-            (config.groups.clone(), config.nodes.clone())
-        };
-        let new_gm = GroupManager::with_alive_set(&groups, &nodes, Some(self.alive_set.clone()));
-        // Migrate runtime choices before wiring callbacks: migration must
-        // not fire persistence or connection interruption.
-        new_gm.migrate_selector_choices_from(&self.group_manager.read());
-        install_interrupt_callback(&new_gm, &self.group_manager, &self.connection_tracker);
-        if let Some(ref db) = self.cache_db {
-            let db_cb = db.clone();
-            new_gm.set_persist_callback(Some(Arc::new(move |group, node| {
-                db_cb.save_selector_choice(group, node);
-            })));
-        }
-        *self.group_manager.write() = Arc::new(new_gm);
-
-        // Refresh health-check registrations and the URLTest idle table to
-        // match the new group membership.
-        let config = self.config.read().await;
-        let (added, removed) = sync_health_check_nodes(&self.alive_set, &config);
-        self.alive_set
-            .sync_urltest_groups(&urltest_group_registrations(&config));
-        self.alive_set
-            .sync_group_check_urls(&group_check_url_registrations(&config));
-        info!(
-            "Group manager rebuilt: {} group(s), health checks +{}/-{} node(s)",
-            config.groups.len(),
-            added,
-            removed,
-        );
-    }
-}
-/// Fields whose current consumers are process-scoped and therefore cannot be
-/// swapped safely by the runtime generation publication. A rejected reload
-/// has not mutated any live state.
-use super::reload_policy::restart_required_changes;
-
-use super::reload_warm::*;
-
-pub(super) use super::reload_subscription::config_with_subscription_nodes;
-
-use super::reload_connectivity::{
-    group_connectivity_snapshot, open_group_connectivity, publish_group_connectivity,
-};
-
-pub(super) use super::reload_connectivity::{
-    build_outbound_id_map, group_datapath_alive, install_interrupt_callback,
-    install_selector_warm_callback,
-};
-
-pub(crate) fn resolve_outbound_nodes(
-    config: &Config,
-    group_manager: &GroupManager,
-    outbound_name: &str,
-    domain: ProbeDomain,
-    ipver: IpVersion,
-) -> Vec<Node> {
-    if let Some(node) = config.builtin_node(outbound_name) {
-        return vec![node];
-    }
-    if let Some(node) = config.nodes.iter().find(|n| n.name == outbound_name) {
-        return vec![node.clone()];
-    }
-    for group in &config.groups {
-        if group.name == outbound_name {
-            let mut nodes =
-                group_manager.select_nodes_in_order_for_domain(&group.name, domain, ipver);
-            // Fallback: IPv6 targets may still be forwarded through nodes that
-            // are only reachable over IPv4 (common for proxy servers with only
-            // an A record). Try IPv4 alive candidates before giving up.
-            if nodes.is_empty() && ipver == IpVersion::V6 {
-                nodes = group_manager.select_nodes_in_order_for_domain(
-                    &group.name,
-                    domain,
-                    IpVersion::V4,
-                );
-                if !nodes.is_empty() {
-                    warn!(
-                        "resolve_outbound_nodes: group '{}' has no IPv6 alive node; falling back to IPv4 alive candidates",
-                        group.name
-                    );
-                }
-            }
-            if nodes.is_empty() {
-                warn!(
-                    "resolve_outbound_nodes: group '{}' has no available node (ipver={:?})",
-                    group.name, ipver
-                );
-                // When all nodes in a group are dead and `final` is configured,
-                // recursively resolve the fallback outbound.
-                if let Some(final_name) = group_manager.get_final_outbound(&group.name) {
-                    info!(
-                        "Group '{}' has no alive nodes, falling back to final outbound '{}'",
-                        group.name, final_name
-                    );
-                    return resolve_outbound_nodes(
-                        config,
-                        group_manager,
-                        &final_name,
-                        domain,
-                        ipver,
-                    );
-                }
-            }
-            return nodes.into_iter().cloned().collect();
-        }
-    }
-    warn!(
-        "Outbound '{}' not found, falling back to direct",
-        outbound_name
-    );
-    vec![Config::builtin_direct_node()]
-}
-
-/// Concrete UDP candidates plus the provenance and IP family selected by
-/// the final outbound resolution. This companion does not change the legacy
-/// TCP/DNS `resolve_outbound_nodes` API.
-#[derive(Debug, Clone)]
-pub(super) struct ResolvedUdpPlan {
-    pub(super) mode: honk_outbound::group::SelectionPlanMode,
-    pub(super) nodes: Vec<Node>,
-    pub(super) ipver: IpVersion,
-}
-
-/// Resolve UDP candidates without inferring policy from candidate count.
-///
-/// A group plan supplies the authoritative/cold provenance directly. Empty
-/// groups may follow `final_outbound`, in which case the terminal outbound's
-/// mode and resolved IP version replace the outer plan. Recursive final
-/// chains are bounded and cycle-safe; a missing final target retains the
-/// historical direct fallback, while a cycle/depth breach fails closed.
-pub(super) fn resolve_udp_outbound_plan(
-    config: &Config,
-    group_manager: &GroupManager,
-    outbound_name: &str,
-    ipver: IpVersion,
-) -> ResolvedUdpPlan {
-    resolve_udp_outbound_plan_inner(
-        config,
-        group_manager,
-        outbound_name,
-        ipver,
-        0,
-        &mut Vec::new(),
-    )
-}
-
-fn resolve_udp_outbound_plan_inner(
-    config: &Config,
-    group_manager: &GroupManager,
-    outbound_name: &str,
-    ipver: IpVersion,
-    depth: usize,
-    visited: &mut Vec<String>,
-) -> ResolvedUdpPlan {
-    if let Some(node) = config.builtin_node(outbound_name) {
-        return ResolvedUdpPlan {
-            mode: honk_outbound::group::SelectionPlanMode::Authoritative,
-            nodes: vec![node],
-            ipver,
-        };
-    }
-    if let Some(node) = config.nodes.iter().find(|node| node.name == outbound_name) {
-        let mut selected_ipver = ipver;
-        let nodes = if group_manager.is_node_selectable_for_domain(
-            node.id,
-            ProbeDomain::DataUdp,
-            selected_ipver,
-        ) {
-            vec![node.clone()]
-        } else if ipver == IpVersion::V6
-            && group_manager.is_node_selectable_for_domain(
-                node.id,
-                ProbeDomain::DataUdp,
-                IpVersion::V4,
-            )
-        {
-            selected_ipver = IpVersion::V4;
-            vec![node.clone()]
-        } else {
-            vec![]
-        };
-        return ResolvedUdpPlan {
-            mode: honk_outbound::group::SelectionPlanMode::Authoritative,
-            nodes,
-            ipver: selected_ipver,
-        };
-    }
-    let Some(group) = config
-        .groups
-        .iter()
-        .find(|group| group.name == outbound_name)
-    else {
-        warn!(
-            "UDP outbound '{}' not found, falling back to direct",
-            outbound_name
-        );
-        return ResolvedUdpPlan {
-            mode: honk_outbound::group::SelectionPlanMode::Authoritative,
-            nodes: vec![Config::builtin_direct_node()],
-            ipver,
-        };
-    };
-    if depth >= honk_outbound::group::MAX_GROUP_DEPTH
-        || visited.iter().any(|name| name == outbound_name)
-    {
-        warn!(
-            "UDP final outbound resolution for '{}' stopped at recursive cycle/depth",
-            outbound_name
-        );
-        return ResolvedUdpPlan {
-            mode: honk_outbound::group::SelectionPlanMode::Authoritative,
-            nodes: vec![],
-            ipver,
-        };
-    }
-
-    visited.push(outbound_name.to_owned());
-    let mut selected_ipver = ipver;
-    let mut plan =
-        group_manager.selection_plan_for_domain(&group.name, ProbeDomain::DataUdp, selected_ipver);
-    // Proxy servers frequently have only an A record. Preserve that concrete
-    // fallback family for traffic health feedback rather than reporting the
-    // original IPv6 destination family.
-    if plan.nodes.is_empty() && ipver == IpVersion::V6 {
-        plan = group_manager.selection_plan_for_domain(
-            &group.name,
-            ProbeDomain::DataUdp,
-            IpVersion::V4,
-        );
-        if !plan.nodes.is_empty() {
-            selected_ipver = IpVersion::V4;
-            warn!(
-                "UDP group '{}' has no IPv6 alive node; falling back to IPv4 alive candidates",
-                group.name
-            );
-        }
-    }
-    if !plan.nodes.is_empty() {
-        visited.pop();
-        return ResolvedUdpPlan {
-            mode: plan.mode,
-            nodes: plan.nodes.into_iter().cloned().collect(),
-            ipver: selected_ipver,
-        };
-    }
-
-    if let Some(final_name) = group_manager.get_final_outbound(&group.name) {
-        info!(
-            "UDP group '{}' has no available node; falling back to final outbound '{}'",
-            group.name, final_name
-        );
-        let terminal = resolve_udp_outbound_plan_inner(
-            config,
-            group_manager,
-            &final_name,
-            ipver,
-            depth + 1,
-            visited,
-        );
-        visited.pop();
-        return terminal;
-    }
-    visited.pop();
-    ResolvedUdpPlan {
-        mode: plan.mode,
-        nodes: vec![],
-        ipver: selected_ipver,
     }
 }
